ca fait des trucs mais pas les bons

+++ srcs/utils.c
@@ -39,10 +39,10 @@ int	ft_atoi(const char *nptr)
 	return (res * sign);
 }
 
-int	get_time(void)
+long long	get_time(void)
 {
 	struct timeval	time;
 
 	gettimeofday(&time, NULL);
-	return (time.tv_sec * 1000 + time.tv_usec / 1000);
+	return ((long long)(time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
